update (source): Establish chapter 3 dk scripts
Per source notes
 Xinef: Should be in chapter III
We establish chapter 3 script for scarelet enclaved and to make the easy of future development in terms of organization
