+++ src/server/scripts/EasternKingdoms/eastern_kingdoms_script_loader.cpp
@@ -97,6 +97,7 @@ void AddSC_instance_molten_core();
 void AddSC_the_scarlet_enclave();            //Scarlet Enclave
 void AddSC_the_scarlet_enclave_c1();
 void AddSC_the_scarlet_enclave_c2();
+void AddSC_the_scarlet_enclave_c3();
 void AddSC_the_scarlet_enclave_c5();
 void AddSC_instance_scarlet_monastery();     //Scarlet Monastery
 void AddSC_boss_kirtonos_the_herald();
@@ -247,6 +248,7 @@ void AddEasternKingdomsScripts()
     AddSC_the_scarlet_enclave();            //Scarlet Enclave
     AddSC_the_scarlet_enclave_c1();
     AddSC_the_scarlet_enclave_c2();
+    AddSC_the_scarlet_enclave_c3();
     AddSC_the_scarlet_enclave_c5();
     AddSC_instance_scarlet_monastery();     //Scarlet Monastery
     AddSC_boss_kirtonos_the_herald();
